requirements.txt
Django<1.6
dj-database-url==0.2.1
dj-settings-helpers==1.0.0
factory-boy==2.3.1
